+++ lib/src/typedefs/hidden_tabs_menu_item_builder.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../tab_data.dart';
-
-/// Builder for hidden tabs menu item.
-typedef HiddenTabsMenuItemBuilder = Widget Function(
-    BuildContext context, int tabIndex, TabData tabData);